move pubsub to lib folder
